Always return PixelData.num_pixels as double
See https://github.com/pace-neutrons/Horace/issues/272 for bug
description

+++ horace_core/sqw/@PixelData/PixelData.m
@@ -517,7 +517,7 @@ methods
         if isempty(obj.f_accessor_)
             num_pix = size(obj.data, 2);
         else
-            num_pix = obj.f_accessor_.npixels;
+            num_pix = double(obj.f_accessor_.npixels);
         end
     end
